refactor(ios): isolate snapshot visibility fold
